Fix tx infinite loading
