+++ packages/ui/src/hooks/dapp/use-transactions.ts
@@ -62,7 +62,11 @@ export const useTransactions = (addresses: string[]) => {
 									affected_global_entities_filter: [address],
 									limit_per_page: 5,
 									cursor: pageParam?.[idx].next_cursor || null,
-									at_ledger_state: pageParam?.[idx].next_cursor ? pageParam?.[idx].ledger_state : null,
+									at_ledger_state: pageParam?.[idx].next_cursor
+										? {
+												state_version: pageParam?.[idx].ledger_state.state_version,
+										  }
+										: null,
 								},
 						  }),
 				),
